requirements.txt
markdown2==2.3.8
python-slugify==4.0.0
text-unidecode==1.3
